added search and controllers for Author and Genre deleted WeatherForecast

+++ BookManagement/Controllers/BookController.cs
@@ -4,6 +4,8 @@ using BookManagement.Models;
 using BookManagement.Context;
 using Microsoft.EntityFrameworkCore;
 using System.Threading.Tasks;
+using System;
+using System.Linq;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -20,51 +22,86 @@ namespace BookManagement.Controllers
             _context = context;
         }
 
-        // GET: api/<ValuesController>
+        // GET: /Book
         [HttpGet]
-        public async Task<IEnumerable<Book>> Get()
+        public async Task<ActionResult<IEnumerable<Book>>> GetBooks()
         {
             return await _context.Books
-                .Include(b => b.Genre)
-                .Include(b => b.Author)
-                .AsNoTracking()
-                .ToListAsync();
+                        .Include(b => b.Genre)
+                        .Include(b => b.Author)
+                        .AsNoTracking()
+                        .ToListAsync();
         } 
 
-        // GET api/<ValuesController>/5
+        // GET /Book/5
         [HttpGet("{id}")]
-        public async Task<Book> Get(int id)
+        public async Task<ActionResult<Book>> GetBook(int id)
         {
             return await _context.Books
-                .Include(b => b.Genre)
-                .Include(b => b.Author)
-                .AsNoTracking()
-                .FirstOrDefaultAsync(b => b.BookID == id);
+                        .Include(b => b.Genre)
+                        .Include(b => b.Author)
+                        .AsNoTracking()
+                        .FirstOrDefaultAsync(b => b.BookID == id);
         }
 
-        // POST api/<ValuesController>
+        // POST /Book
         [HttpPost]
-        public void Post([FromBody] Book value)
+        public async Task<ActionResult<Book>> PostBook([FromBody] Book book)
         {
-            _context.Books.Add(value);
-            _context.SaveChanges();
+            _context.Books.Add(book);
+            await _context.SaveChangesAsync();
+
+            return CreatedAtAction(nameof(GetBook), new { id = book.BookID }, book);
         }
 
-        // PUT api/<ValuesController>/5
+        // PUT /Book/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] Book value)
+        public async Task<IActionResult> PutBook(int id, [FromBody] Book book)
         {
-            _context.Books.Update(value);
-            _context.SaveChanges();
+            if(id != book.BookID)
+            {
+                return BadRequest();
+            }
+
+            _context.Entry(book).State = EntityState.Modified;
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if(!BookExists(id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+            return NoContent();
         }
 
-        // DELETE api/<ValuesController>/5
+        // DELETE /Book/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public async Task<IActionResult> DeleteBook(int id)
         {
-            var book = _context.Books.Find(id);
+            var book = await _context.Books.FindAsync(id);
+            if(book == null)
+            {
+                return NotFound();
+            }
+
             _context.Books.Remove(book);
-            _context.SaveChanges();
+            await _context.SaveChangesAsync();
+
+            return NoContent();
+        }
+
+        private bool BookExists(int id)
+        {
+            return _context.Books.Any(b => b.BookID == id);
         }
     }
 }
